Implementação do hook useFile
