+++ src/components/list-files/list-files.tsx
@@ -1,4 +1,4 @@
-import { MouseEvent, RefObject } from 'react'
+import { RefObject } from 'react'
 import {
   ListFilesStyled,
   ItemFilesStyled,
@@ -6,8 +6,6 @@ import {
   ButtonFileStyled,
   TextLinkStyled,
 } from './list-files-styled'
-
-import * as FileActions from 'common/file-actions'
 
 import { File, Status } from 'resources/files/types'
 import * as Icon from 'ui/icons'
@@ -82,34 +80,7 @@ function ItemFiles({
   refInputFileName,
   refEditorTextArea,
 }: ItemFilesProps) {
-  const { removeFile } = useFile()
-
-  const handleClick = (e: MouseEvent) => {
-    e.preventDefault()
-    setCurrentFileId(fileId)
-
-    const filesNew = files.map((file) => {
-      file.active = file.id === fileId
-      file.status = file.id === fileId ? 'editing' : 'saved'
-
-      return file
-    })
-
-    setFiles(filesNew)
-
-    FileActions.setFileList(filesNew)
-
-    if (refInputFileName.current) {
-      refInputFileName.current.value = fileName
-      refInputFileName.current.focus()
-    }
-
-    if (refEditorTextArea.current) {
-      refEditorTextArea.current.value = fileContent
-
-      setMkdText(fileContent)
-    }
-  }
+  const { onSelected, removeFile } = useFile()
 
   return (
     <ItemFilesStyled
@@ -120,7 +91,19 @@ function ItemFiles({
       <LinkFilesStyled
         href={`/file/${fileId}`}
         onClick={(e) => {
-          handleClick(e)
+          e.preventDefault()
+          setCurrentFileId(fileId)
+
+          onSelected({
+            files,
+            fileId,
+            fileName,
+            fileContent,
+            refInputFileName,
+            refEditorTextArea,
+            setFiles,
+            setMkdText,
+          })
         }}
       >
         {!fileActive ? <Icon.File /> : <Icon.FileActive />}
