+++ src/hooks/useFile.ts
@@ -25,8 +25,52 @@ type RemoveFileType = {
   setFiles: Function
 }
 
+type OnSelectedType = {
+  files: File[]
+  fileId: string
+  fileName: string
+  fileContent: string
+  refInputFileName: RefObject<HTMLInputElement>
+  refEditorTextArea: RefObject<HTMLTextAreaElement>
+  setFiles: Function
+  setMkdText: Function
+}
+
 export function useFile() {
   const keyApp = 'markee-app'
+
+  const onSelected = ({
+    files,
+    fileId,
+    fileName,
+    fileContent,
+    refInputFileName,
+    refEditorTextArea,
+    setFiles,
+    setMkdText,
+  }: OnSelectedType) => {
+    const filesNew = files.map((file) => {
+      file.active = file.id === fileId
+      file.status = file.id === fileId ? 'editing' : 'saved'
+
+      return file
+    })
+
+    setFiles(filesNew)
+
+    localForage.setItem(keyApp, JSON.stringify(filesNew))
+
+    if (refInputFileName.current) {
+      refInputFileName.current.value = fileName
+      refInputFileName.current.focus()
+    }
+
+    if (refEditorTextArea.current) {
+      refEditorTextArea.current.value = fileContent
+
+      setMkdText(fileContent)
+    }
+  }
 
   const addFile = ({
     files,
@@ -119,11 +163,6 @@ export function useFile() {
     })
   }
 
-  function getFiles() {
-    if (!localForage.getItem(keyApp)) return
-    return localForage.getItem(keyApp)
-  }
-
   function removeFile({ files, fileId, setFiles }: RemoveFileType) {
     const filesNew = files.filter((file) => file.id !== fileId)
 
@@ -137,6 +176,6 @@ export function useFile() {
     editFiles,
     saveFiles,
     removeFile,
-    getFiles,
+    onSelected,
   }
 }
